backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.97.0
anyio==4.12.1
cachetools==6.2.6
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.1.8
cryptography==47.0.0
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.18.0
exceptiongroup==1.3.1
fastapi==0.128.8
fsspec==2025.10.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.13
jiter==0.14.0
markdown-it-py==3.0.0
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.1
openai==2.32.0
packaging==26.2
postgrest==2.29.0
propcache==0.4.1
pycparser==2.23
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
pyiceberg==0.10.0
PyJWT==2.12.1
pyparsing==3.3.2
pyroaring==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
realtime==2.29.0
requests==2.32.5
rich==14.3.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.49.3
storage3==2.29.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.29.0
supabase-auth==2.29.0
supabase-functions==2.29.0
tenacity==9.1.2
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.39.0
websockets==15.0.1
yarl==1.22.0
